[REFACTOR] 장바구니 병합 오류 해결
