+++ WhyNComing/src/main/java/org/sparta/whyncoming/store/domain/entity/Store.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import org.sparta.whyncoming.common.entity.BaseActorEntity;
 import org.sparta.whyncoming.order.domain.entity.Order;
 import org.sparta.whyncoming.order.domain.entity.Review;
-import org.sparta.whyncoming.product.domain.entity.Cart;
+import org.sparta.whyncoming.order.domain.entity.Cart;
 import org.sparta.whyncoming.product.domain.entity.Product;
 import org.sparta.whyncoming.user.domain.entity.User;
